feat: rewrite as native Kotlin + Jetpack Compose Android app
Complete rewrite of the Balam CRM app from Capacitor/Next.js to native
Android using Kotlin + Jetpack Compose + Hilt + Retrofit. Removes all
Expo/React Native/web files and replaces with a production-ready Android
project targeting API 35 with minSdk 26.

Implements all 19 screens: Login, Register, Dashboard, Policies,
PolicyDetail, FUP, Families, FamilyDetail, Clients, ClientDetail,
Commission, Loans, SB, Leads, Activities, GST, Reports, Profile, More.

Architecture: MVVM with Hilt DI, StateFlow UiState, Retrofit2/OkHttp3,
EncryptedSharedPreferences token storage, Material3 theme.

+++ app/src/main/java/com/balam/crm/viewmodel/ViewModels.kt
@@ -545,7 +545,7 @@ class GSTViewModel @Inject constructor(private val api: ApiService) : ViewModel(
     private val _gst = MutableStateFlow<UiState<GSTCalculation>>(UiState.Idle)
     val gst: StateFlow<UiState<GSTCalculation>> = _gst
 
-    fun calculate(policyNo: Long) {
+    fun calculate(policyNo: Int) {
         viewModelScope.launch {
             _gst.value = UiState.Loading
             try {
